📦 refactor: organise into nuxt 4 component layers
