+++ apps/web/app/layouts/default.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="min-h-screen w-full m-auto">
+    <slot />
+  </div>
+</template>
